+++ src/main/java/cn/tecnpan/majiang/helloworld/service/QuestionService.java
@@ -10,11 +10,11 @@ public interface QuestionService {
 
     PaginationDto<QuestionDto> getList(Integer pageNo, Integer pageSize);
 
-    PaginationDto<QuestionDto> list(Integer id, Integer pageNo, Integer pageSize);
+    PaginationDto<QuestionDto> list(Long id, Integer pageNo, Integer pageSize);
 
-    QuestionDto getById(Integer id);
+    QuestionDto getById(Long id);
 
     void createOrUpdate(Question question);
 
-    void incView(Integer id);
+    void incView(Long id);
 }
